extract ReportNotFoundGuard and LinkedActionNotFoundGuard from ReportScreen
